+++ pulsar/apps/tasks/worker.py
@@ -27,15 +27,29 @@ This is an interface for using the same IOLoop class of other workers.'''
 
 class TaskScheduler(pulsar.WorkerMonitor):
     
-    def actor_addtask(self, task_name, targs, tkwargs, **kwargs):
-        request = self.app.make_request(task_name, targs, tkwargs, ack=True, **kwargs)
-        self.task_queue.put(request)
-        return request
+    def _addtask(self, caller, task_name, targs, tkwargs, ack = True, **kwargs):
+        try:
+            request = self.app.make_request(task_name, targs, tkwargs, **kwargs)
+            self.task_queue.put((None,request))
+            if ack:
+                return request
+        except Exception as e:
+            return e
+    
+    def actor_addtask(self, caller, task_name, targs, tkwargs, ack=True, **kwargs):
+        return self._addtask(caller, task_name, targs, tkwargs, ack = True, **kwargs)
         
-    def actor_addtask_noack(self, task_name, targs, tkwargs, **kwargs):
-        request = self.app.make_request(task_name, targs, tkwargs, **kwargs)
-        self.task_queue.put(request)
+    def actor_addtask_noack(self, caller, task_name, targs, tkwargs, ack=False, **kwargs):
+        return self._addtask(caller, task_name, targs, tkwargs, ack = False, **kwargs)
     actor_addtask_noack.ack = False
+    
+    def actor_task_finished(self, caller, response):
+        self.app.task_finished(response)
+    actor_task_finished.ack = False
+    
+    def actor_get_task(self, caller, id):
+        return self.app.get_task(id)
+    actor_get_task.ack = True
         
 
 class Worker(pulsar.Worker):