+++ pulsar/http/__init__.py
@@ -2,6 +2,9 @@ from pulsar.utils.importer import import_module
 
 from .client import HttpClient
 
+EMPTY_TUPLE = ()
+EMPTY_DICT = {}
+
 def get_httplib(cfg = None):
     name = None if not cfg else cfg.settings['httplib'].value
     if name == 'gunicorn':
@@ -10,17 +13,36 @@ def get_httplib(cfg = None):
         return import_module('pulsar.http.base')
     
     
-def queueTask(taskname, doc = '', ack = True, taskqueue = "taskqueue"):
+
+def actorCall(function, doc = '', ack = True, server = "taskqueue"):
+    
+    def _(self, request, **kwargs):
+        worker = request.environ['pulsar.worker']
+        tk = worker.ACTOR_LINKS[server]
+        if ack:
+            return tk.send(worker.aid, (EMPTY_TUPLE,kwargs), name=function, ack=True)
+        else:
+            tk.send(worker.aid, (EMPTY_TUPLE,kwargs), name='addtask_noack', ack=False)
+        
+    _.__doc__ = doc
+    _.__name__ = function
+    return _
+    
+        
+def queueTask(taskname, doc = '', ack = True, server = "taskqueue"):
     # A decorator for running a taskname in the taskqueue
     
     def _(self, request, **kwargs):
         worker = request.environ['pulsar.worker']
-        tk = worker.ACTOR_LINKS[taskqueue]
-        r = tk.send(worker.aid, ((taskname,),kwargs), 'addtask')
+        tk = worker.ACTOR_LINKS[server]
+        args = (taskname,(),kwargs)
         if ack:
-            return r
+            return tk.send(worker.aid, (args,EMPTY_DICT), name='addtask', ack=True)
+        else:
+            tk.send(worker.aid, (args,EMPTY_DICT), name='addtask_noack', ack=False)
         
     _.__doc__ = doc
     _.__name__ = taskname
     return _
+