+++ tests/regression/async/tests.py
@@ -1,2 +1,1 @@
-from .actors import *
 from .deferred import *