+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Version 0.1.0 - Development
 
 * First official alpha release.
 * Working for python 2.6 and up, including python 3.
-* Five different type of applications: HTTP server, RPC server, distributed task queue,
+* Five different applications: HTTP server, RPC server, distributed task queue,
   asynchronous test suite and asynchronous shell.
 * 4 examples spanning all applications.
-* **35 regression tests**
+* **38 regression tests**